requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sentence-transformers==2.5.1
pillow==10.2.0
numpy==1.26.4
requests==2.31.0
torch==2.2.1
torchvision==0.17.1
python-multipart==0.0.9
llama-cpp-python==0.2.90
